Greedy
In this problem an approach is used such that the coin changed from even to even   and odd to odd the cost is zero but the cost to move from even to odd or vice versa is  one. So  Check coins in even position and odd position and return the min of two.