Day 3-LinkedList and Simple Math
In merge the two linked list compare and traverse the nodes in two linked list simultaneously and make a new based on the comparison. In perfect Number i applied brute force to check all the divisor and add them then see id their sum is equal to the number but a more efficient way is to make an array of perfect numbers in given range and check weather the number is in given array.