Day 4-LinkedList and Simple Math
In the cycle detection we take two pointers in the form of List Nodes and then we move one pointer faster and one slower if faster pointers next node is the slower one we return true else false also if any of the two pointers are null then No cycle exists. In palindrome I have taken an array with max length  32 according to given limit and then each digit is filled after that from start and end check if the digits are same.