Day 7-LinkedList and Simple Math
Here in even and odd LL we used 4 pointers to make even and odd separate LL  to keep record of the head and tail and then merge them.
In finding column Number in excel we used approach in which we count the number of columns in front of it  and then add them to find the column number.