Day 11-LinkedList and Simple Math
Here In the LL we first calculate the rank of node from the start and then delete it normally. In Factorial of trailing zeroes We calculate the number of twos in the factorial by repeatedly dividing by 2 Till 1 (here repeatedly because some number can have may power of 2 such as 8,24,12)Then do the same for the 5 and return minimum of the two.