Day 13-LinkedList and Simple Math
In remove Duplication of LL we use array to keep track of duplicated values and an hash set to keep track of nodes to be deleted and then delete them accordingly. In product of an array we use approach to see if zero there or not and if not count the negative number in an array to get answer.