Day 24-Queue and Array
In rotate array we use an extra array to copy the values from the rotated point and then recopy to original one. In Implementing stack using queue to push the value we push it in empty queue and then push all values of the another queue and then switch them and rest of the operations are same as that of a queue.