Day 27-Binary Tree and Array
In path sum we check only for root node else go to left and right sub-trees for verification. In Maximum Subarray if all numbers negative then select maximum of array else add the temp to current value and do it repeatedly till the temp value is less than zero then reset it to zero and continue.