41-Binary Search Tree and Strings
In BST if only one child node of deleted node then it is easy to replace if two then find smallest number larger than deleted node or greatest number smaller than the deleted node to replace it. In Finding Largest Prefix first find for two then compare that prefix with rest of the  strings to get appropriate answer.