42-Binary Search Tree and Strings
In a BST if we have to validate it we have to check the min and max value that a tree node can have if not in range return false. In reverse the words of String we use stack to store the words and then pop it sequentially to get words in a desired order.