Day 47-Strings and Bit Manipulation
In this Number of bits we find bits in similar way in which we convert a decimal number to binary number. In Counting Number of Sub String we get all three character from end we then go on till we have all three character from shortening it from start then repeat both steps.