Day 48-Arrays and Bit Manipulation
In check if it is a good array or not we used the Bézout's Identity in which the GCD of array is minimum difference so wee need the GCD of array to be 1. In hamming Distance problem we use XOR to get number of different  bits and if a bit is set just count them to get hamming distance.